factor theme out to separate module
